Add upgrades and setup input
Add functionality for upgrade levels and passing of next upgrade. This passes the next upgrade to the list of upgrades once a upgrade is chosen. This also uses a Fisher-Yates algorithm to get the upgrades that will be displayed to be chosen. This allows for no two of the same upgrades to be displayed.